requirements.txt
anthropic>=0.84.0
fastapi>=0.135.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
python-docx>=1.1.0
